requirements.txt
streamlit      
langchain
openai
pinecone
PyPDF2
python-dotenv
tiktoken
unstructured
numpy
pandas
python-multipart
sentence-transformers
langchain-community
langchain-openai